+++ AzureActiveDirectory/azure_ad/get_sign_ins.py
@@ -1,6 +1,11 @@
 from pydantic import BaseModel
+import asyncio
 
-from .base import MicrosoftGraphAction, SingleUserArguments
+from .base import MicrosoftGraphAction, SingleUserArguments, IdArguments
+
+from kiota_abstractions.native_response_handler import NativeResponseHandler
+from kiota_http.middleware.options import ResponseHandlerOption
+from msgraph.generated.users.item.messages.messages_request_builder import MessagesRequestBuilder
 
 
 class DeviceDetail(BaseModel):
@@ -100,21 +105,43 @@ class GetSignInsResults(BaseModel):
 
 class GetSignInsAction(MicrosoftGraphAction):
     name = "Get SignIns"
-    description = "Get the last sign ins of an Azure AD user. Requires the AuditLog.Read.All and Directory.Read.All permissions."  # noqa: E501
-    results_model = GetSignInsResults
+    description = "Get the last sign ins of an Azure AD user. Requires the AuditLog.Read.All and Directory.Read.All app permissions."  # noqa: E501
+    # results_model doesn't support async model
+    # results_model = GetSignInsResults
 
-    def run(self, arguments: SingleUserArguments):
-        params = {}
+    async def query_get_user_signin(self, req_conf):
+        return await self.client.audit_logs.sign_ins.get(request_configuration=req_conf)
 
-        if arguments.id:
-            params["$filter"] = f"userId eq '{arguments.id}'"
-        elif arguments.userPrincipalName:
-            params["$filter"] = f"userPrincipalName eq '{arguments.userPrincipalName}'"
-
-        response = self.client.get(
-            "/auditLogs/signIns",
-            params=params,
+    async def run(self, arguments: IdArguments):
+        query_params = MessagesRequestBuilder.MessagesRequestBuilderGetQueryParameters(
+            filter=f"userId eq '{arguments.id}'"
         )
+        request_configuration = MessagesRequestBuilder.MessagesRequestBuilderGetRequestConfiguration(
+            options=[ResponseHandlerOption(NativeResponseHandler())], query_parameters=query_params
+        )
+
+        response = await self.query_get_user_signin(request_configuration)
+
         response.raise_for_status()
 
         return {"signIns": response.json()["value"]}
+
+
+class RevokeSignInsSessionsAction(MicrosoftGraphAction):
+    name = "Revoke SignIns Sessions"
+    description = "Invalidates all the refresh tokens issued to applications for a user. Requires the User.ReadWrite.All or Directory.ReadWrite.All permissions."  # noqa: E501
+
+    async def query_revoke_signin(self, signIn_param, req_conf):
+        return await self.client.users.by_user_id(signIn_param).revoke_sign_in_sessions.post(
+            request_configuration=req_conf
+        )
+
+    async def run(self, arguments: SingleUserArguments):
+        request_configuration = MessagesRequestBuilder.MessagesRequestBuilderGetRequestConfiguration(
+            options=[ResponseHandlerOption(NativeResponseHandler())],
+        )
+        signIn_param = arguments.id or arguments.userPrincipalName
+
+        response = await self.query_revoke_signin(signIn_param, request_configuration)
+
+        response.raise_for_status()