+++ Darktrace/darktrace_modules/client/__init__.py
@@ -12,8 +12,13 @@ class ApiClient(requests.Session):
         private_key: str,
         nb_retries: int = 5,
         ratelimit_per_minute: int = 20,
+        default_headers: dict[str, str] | None = None,
     ):
         super().__init__()
+
+        if default_headers is not None:
+            self.headers.update(default_headers)
+
         self.auth = ApiKeyAuthentication(public_key, private_key)
         self.mount(
             "https://",