+++ Office365/office365/metrics.py
@@ -18,7 +18,7 @@ FORWARD_EVENTS_DURATION = Histogram(
 )
 
 EVENTS_LAG = Histogram(
-    name="event_lags",
+    name="events_lags",
     documentation="The delay, in seconds, from the date of the last event",
     namespace=prom_namespace,
     labelnames=["intake_key"],