+++ CatoNetwork/CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### 2023-11-08 - 1.2.2
+
+#### Fixed
+
+- fix metrics
+
+### 2023-10-27 - 1.2.1
+
+#### Changed
+
+- remove BETA flag
+
+### 2023-10-25 - 1.2
+
+#### Changed
+
+- Fix run method
+
+### 2023-10-25 - 1.1
+
+#### Changed
+
+- update dependencies
+
 ### 2023-09-13 - 1.0.2
 
 #### Changed